Added support for promises to mesh.init
When calling mesh.init, you can pass in a promise as an argument. If a promise is passed in, then your init and draw methods will not be called until the promise resolves.

This allows you to call async methods in setup, and use information from mesh, and not have your init and draw methods called until your async methods have returned. For example, you may need to load some image / pixel data in your initialization and have that avaliable before init and draw is called.