Refactor mouse events
Big refactor of mouse events, and implimentation for new events. Can now listen for mouse events by calling:

mesh.listen();

passing in the event handler (named depending on the event you want to listen to). For example, for click events:

mesh.listen(click);

or to register for multiple events at a time:

mesh.listen(click, mouseup, mousedown, mousemove);

to stop listening for an event, just pass in false as the last argument:

mesh.listen(mousemove, false);

Note, all events are dispatched in sync with browser drawing (onAnimationFrameusing requestAnimationFrame), and will be dispatched before draw() is called.

Currently, events are dispatched in the order the listeners are added. Ill fix this to have an explicit event order.